backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.4.2
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.0.0
httpx==0.25.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
PyJWT==2.12.1
